+++ packages/ui/src/api/strava.ts
@@ -1,6 +1,9 @@
 import { StravaActivity } from '@torq/strava-api';
 
+import env from '@/env';
+
 import { apiRequest } from './client';
+import { ENDPOINTS } from './constants';
 
 /**
  * Initiate OAuth flow by redirecting to backend auth endpoint.
@@ -8,7 +11,7 @@ import { apiRequest } from './client';
  * @returns {void}
  */
 export function authorizeStrava(): void {
-  window.location.href = '/strava/auth';
+  window.location.href = `${env.apiUrl}${ENDPOINTS.STRAVA_AUTH}`;
 }
 
 /**
@@ -16,23 +19,23 @@ export function authorizeStrava(): void {
  * @returns {Promise<StravaActivity[]>} Array of user activities
  */
 export async function fetchActivities(): Promise<StravaActivity[]> {
-  return apiRequest<StravaActivity[]>('/strava/activities');
+  return apiRequest<StravaActivity[]>(ENDPOINTS.STRAVA_ACTIVITIES);
 }
 
 /**
  * Fetch specific activity by ID.
- * @param {number} id - Activity ID
+ * @param {string} id - Activity ID
  * @returns {Promise<StravaActivity>} Activity data
  */
-export async function fetchActivity(id: number): Promise<StravaActivity> {
-  return apiRequest<StravaActivity>(`/strava/activity/${id}`);
+export async function fetchActivity(id: string): Promise<StravaActivity> {
+  return apiRequest<StravaActivity>(ENDPOINTS.STRAVA_ACTIVITY(id));
 }
 
 /**
  * Fetch specific activity signals by activity ID.
- * @param {number} id - Activity ID.
+ * @param {string} id - Activity ID.
  * @returns {Promise<StravaActivity>} Activity signals.
  */
-export async function fetchActivitySignals(id: number): Promise<StravaActivity> {
-  return apiRequest<StravaActivity>(`/strava/activities/${id}/signals`);
+export async function fetchActivitySignals(id: string): Promise<StravaActivity> {
+  return apiRequest<StravaActivity>(ENDPOINTS.STRAVA_ACTIVITY_SIGNALS(id));
 }